Marker and SelectionController are simplificated. Fixed marker moving formulas. Added marker rotation. Refactoring.
